+++ CHANGELOG.md
@@ -5,7 +5,6 @@
 ### Added
 
 - Added borehole geometry panel.
-- Added polygon filter for boreholes.
 
 ### Changed
 
@@ -17,6 +16,7 @@
 - Removed filter by map functionality.
 - Moved icon to handle custom layers to sidebar.
 - Show more entries on the boreholes table when screensize is smaller than 1200px.
+- Calculate TVD values in General panel using the geometry of the borehole.
 
 ### Fixed